pyproject.toml
[project]
name = "financial-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.13.2",
    "dagster-postgres>=0.29.2",
    "dagster-webserver>=1.13.2",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "respx>=0.23.1",
    "sqlalchemy>=2.0.49",
    "yfinance>=1.3.0",
]
